// src/common/types/access.ts
export enum OmsAccessCodeSuffix {
  /** Oms系统 */
  System = '',
  /** 行情模块 */
  Mkt = 'Mkt',
  /** 利率行情菜单 */
  MktMenu = 'MktMenu',
  /** 行情页面 */
  MktPage = 'MktPage',
  /** 报价及报价编辑 */
  MktQuote = 'MktQuote',
  /** 市场成交 */
  MktDeal = 'MktDeal',
  /** 债券详情 */
  MktBond = 'MktBond',
  /** 行情日志 */
  MktLog = 'MktLog',
  /** NCDP打开报表 */
  Report = 'Report',
  /** 协同报价模块 */
  Collaboration = 'Collaboration',
  /** 协同报价 */
  CollaborationMenu = 'CollaborationMenu',
  /** iQuote模块 */
  IQuote = 'IQuote',
  /** iQuote菜单 */
  IQuoteMenu = 'IQuoteMenu',
  /** 行情追踪模块 */
  Trace = 'Trace',
  /** 行情追踪菜单 */
  TraceMenu = 'TraceMenu',
  /** 计算器模块 */
  Cal = 'Cal',
  /** 计算器菜单 */
  CalMenu = 'CalMenu',
  /** 点价模块 */
  SpotPricing = 'SpotPricing',
  /** 点价菜单 */
  SpotPricingMenu = 'SpotPricingMenu',
  /** 点价板、成交记录 */
  SpotPricingPage = 'SpotPricingPage',
  /** 点价操作 */
  SpotPricingAction = 'SpotPricingAction',
  /** 点价refer */
  SpotPricingRefer = 'SpotPricingRefer',
  /** 成交记录克隆、编辑、删除 */
  SpotPricingRecordEdit = 'SpotPricingRecordEdit',
  /** 成交记录发送 */
  SpotPricingRecordSend = 'SpotPricingRecordSend',
  /** 成交记录操作日志 */
  SpotPricingRecordLog = 'SpotPricingRecordLog',
  /** 历史记录 */
  SpotPricingHistPage = 'SpotPricingHistPage',
  /** 历史记录-本人 */
  SpotPricingHistSelf = 'SpotPricingHistSelf',
  /** 历史记录-全部 */
  SpotPricingHistAll = 'SpotPricingHistAll',
  /** 明细模块 */
  DealDetail = 'DealDetail',
  /** 明细菜单 */
  DealDetailMenu = 'DealDetailMenu',
  /** 明细页面 */
  DealDetailPage = 'DealDetailPage',
  /** 桥信息编辑 */
  DealDetailBridgeEdit = 'DealDetailBridgeEdit',
  /** 发单信息编辑 */
  DealDetailSendEdit = 'DealDetailSendEdit',
  /** 明细-发送 */
  DealDetailSend = 'DealDetailSend',
  /** 明细-催单 */
  DealDetailBuzz = 'DealDetailBuzz',
  /** 新建桥机构 */
  DealDetailBridgeInstAdd = 'DealDetailBridgeInstAdd',
  /** 明细-操作日志 */
  DealDetailLog = 'DealDetailLog',
  /** 代付机构费率维护 */
  DealDetailPayforRate = 'DealDetailPayforRate',
  /** 过桥模块 */
  Bridge = 'Bridge',
  /** 过桥菜单 */
  BridgeMenu = 'BridgeMenu',
  /** 过桥页面 */
  BridgePage = 'BridgePage',
  /** 过桥-本人 */
  BridgeSelf = 'BridgeSelf',
  /** 过桥-全部 */
  BridgeAll = 'BridgeAll',
  /** 桥机构添加、删除、修改 */
  BridgeInstEdit = 'BridgeInstEdit',
  /** 桥信息编辑 */
  BridgeRecordEdit = 'BridgeRecordEdit',
  /** 过桥-操作日志 */
  BridgeLog = 'BridgeLog',
  /** 成交单模块 */
  ReceiptDeal = 'ReceiptDeal',
  /** 成交单菜单 */
  ReceiptDealMenu = 'ReceiptDealMenu',
  /** 成交单页面 */
  ReceiptDealPage = 'ReceiptDealPage',
  /** 成交单-本人 */
  ReceiptDealSelf = 'ReceiptDealSelf',
  /** 成交单-全部 */
  ReceiptDealAll = 'ReceiptDealAll',
  /** 成交录入、编辑 */
  ReceiptDealEdit = 'ReceiptDealEdit',
  /** 成交确认、提交、毁单 */
  ReceiptDealSubmit = 'ReceiptDealSubmit',
  /** 成交单-操作日志 */
  ReceiptDealLog = 'ReceiptDealLog',
  /** 设置模块 */
  Setting = 'Setting',
  /** 设置菜单 */
  SettingMenu = 'SettingMenu',
  /** 指定授权人设置 */
  SettingAssign = 'SettingAssign'
}